+++ stable_baselines3/common/utils.py
@@ -575,7 +575,7 @@ def obs_as_tensor(obs: Union[np.ndarray, dict[str, np.ndarray]], device: th.devi
         return {key: th.as_tensor(_obs, device=device) for (key, _obs) in obs.items()}
     else:
         raise Exception(f"Unrecognized type of observation {type(obs)}")
-
+        return obs
 
 def should_collect_more_steps(
     train_freq: TrainFreq,